% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steponeR.R
\name{steponeR}
\alias{steponeR}
\title{Import StepOne™ qPCR data}
\usage{
steponeR(files, delim = ",", target.ratios, fluor.norm, copy.number, ploidy,
  extract)
}
\arguments{
\item{files}{A list of data files to import (e.g., generated by \code{list.files})}

\item{delim}{The delimiter used in the data files}

\item{target.ratios}{(optional) A list of target ratios to calculate, each
specified as two target names separated by a period (e.g., "C.D"). Target
names must match those in the "Target Name" column of exported data.}

\item{fluor.norm}{(optional) A list of values to subtract from cycle
threshold values for each target, e.g. to normalize for differences in
fluorescence intensity among targets. Target names must match those in the
"Target Name" column of exported data.}

\item{copy.number}{(optional) A list of the number of copies of the target
locus per cell, used to convert target ratios to cell ratios. Target names
must match those in the "Target Name" column of exported data.}

\item{ploidy}{(optional) A list of the ploidy of each target, used to convert
target ratios to cell ratios. Target names must match those in the "Target
Name" column of exported data.}

\item{extract}{(optional) A list of DNA extraction efficiency values for each
target, used to convert target ratios to cell ratios. Target names must
match those in the "Target Name" column of exported data.}
}
\value{
A list with the following named elements:
  \describe{
    \item{unknowns}{The qPCR data (i.e., CT values) for each target in each sample.}
    \item{result}{The target ratios calculated (if any) for each sample}
  }
}
\description{
This function imports qPCR data exported from StepOne™ Software and
optionally calculates target quantities based on a standard curve, or
calculates target ratios.
}
\examples{
# Example data files provided with steponeR package)
file1 <- system.file("extdata/steponedata_1.csv", package = "steponeR")
file2 <- system.file("extdata/steponedata_2.csv", package = "steponeR")

# Load data from files and calculate target ratios
files <- list(file1, file2)
df <- steponeR(files=files, target.ratios=c("C.Mcap", "D.Mcap"),
               fluor.norm=list(C=2.26827, D=0, Mcap=0.84815),
               copy.number=list(C=10, D=2, Mcap=1),
               ploidy=list(C=1, D=1, Mcap=2),
               extract=list(C=0.813, D=0.813, Mcap=0.982))

# Raw data for unknowns
head(df$unknowns)
# Target ratio results
head(df$result)
}
\references{
For more information about the implementation of these calculations and the
use of each parameter in calculating cell ratios from qPCR data, see:

    Cunning R, Baker AC (2013) Excess algal symbionts increase the susceptibility of reef
corals to climate change. Nature Clim Change 3:259-262.
\url{http://dx.doi.org/10.1038/nclimate1711}

    Mieog JC, van Oppen MJH, Berkelmans R, Stam WT, Olsen JL (2009)
Quantification of algal endosymbionts (*Symbiodinium*) in coral tissue using
real-time PCR. Mol Ecol Resour 9:74–82.
\url{http://dx.doi.org/10.1111/j.1755-0998.2008.02222.x}
}
